Feature/chat

+++ src/controllers/TopicController.php
@@ -4,7 +4,7 @@
  * @Email: steven@sbandaogo.com
  * @Date: 2023-09-09 14:50:14 
  * @Last Modified by: Steven Bandaogo
- * @Last Modified time: 2023-09-09 18:09:47
+ * @Last Modified time: 2023-09-11 23:36:51
  * @Description: Manages topics, including creating and deleting topics.
  */
 
@@ -50,7 +50,7 @@ class TopicController
      */
     public static function delete()
     {
-        if (isset($_GET['id'])) {
+        if (isset($_GET['id']) && isset($_GET['id'])) {
             /** Topic deletion */
             Topic::delete(['id' => $_GET['id']]);
 
@@ -61,5 +61,30 @@ class TopicController
         /** Redirection */
         header("location:" . BASE);
         exit();
+    }
+
+    /**
+     * Display the chat page of the selected topic and manage message
+     *  sending.
+     */
+    public static function chat()
+    {
+
+        //* removing every errors saved of the current session.
+        unset($_SESSION['error']);
+
+        //* if topic has been correctly transmitted
+        if (isset($_GET['id']) && !empty($_GET['id'])) {
+            /** Message sending */
+            if (!empty($_POST)) {
+                MessageController::add_message();
+            }
+            /** Get topic information */
+            $topic = Topic::findById(['id' => $_GET['id']]);
+            /** Messages listing */
+            $messages = Message::findByTopic(['id_topic' => $_GET['id']]);
+
+            include(VIEWS . "app/chat.php");
+        }
     }
 }
